+++ packages/lisk-transactions/src/0_transfer.ts
@@ -12,27 +12,13 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-import * as cryptography from '@liskhq/lisk-cryptography';
-import { BYTESIZES, TRANSFER_FEE } from './constants';
+import { getAddressFromPublicKey } from '@liskhq/lisk-cryptography';
+import { TRANSFER_FEE } from './constants';
 import {
-	PartialTransaction,
-	TransferAsset,
+	createBaseTransaction,
 	TransferTransaction,
-} from './transaction_types';
-import {
-	prepareTransaction,
-	validateAddress,
-	validatePublicKey,
-	validateTransferAmount,
-} from './utils';
-
-const createAsset = (data?: string): TransferAsset => {
-	if (data && data.length > 0) {
-		return { data };
-	}
-
-	return {};
-};
+	validateInputs,
+} from './transactions';
 
 export interface TransferInputs {
 	readonly amount: string;
@@ -44,51 +30,7 @@ export interface TransferInputs {
 	readonly timeOffset?: number;
 }
 
-const validateInputs = ({
-	amount,
-	recipientId,
-	recipientPublicKey,
-	data,
-}: TransferInputs): void => {
-	if (!validateTransferAmount(amount)) {
-		throw new Error('Amount must be a valid number in string format.');
-	}
-
-	if (!recipientId && !recipientPublicKey) {
-		throw new Error(
-			'Either recipientId or recipientPublicKey must be provided.',
-		);
-	}
-
-	if (typeof recipientId !== 'undefined') {
-		validateAddress(recipientId);
-	}
-
-	if (typeof recipientPublicKey !== 'undefined') {
-		validatePublicKey(recipientPublicKey);
-	}
-
-	if (
-		recipientId &&
-		recipientPublicKey &&
-		recipientId !== cryptography.getAddressFromPublicKey(recipientPublicKey)
-	) {
-		throw new Error('recipientId does not match recipientPublicKey.');
-	}
-
-	if (data && data.length > 0) {
-		if (typeof data !== 'string') {
-			throw new Error(
-				'Invalid encoding in transaction data. Data must be utf-8 encoded string.',
-			);
-		}
-		if (data.length > BYTESIZES.DATA) {
-			throw new Error('Transaction data field cannot exceed 64 bytes.');
-		}
-	}
-};
-
-export const transfer = (inputs: TransferInputs): TransferTransaction => {
+export const transfer = (inputs: TransferInputs): object => {
 	validateInputs(inputs);
 	const {
 		data,
@@ -96,29 +38,40 @@ export const transfer = (inputs: TransferInputs): TransferTransaction => {
 		recipientPublicKey,
 		passphrase,
 		secondPassphrase,
-		timeOffset,
 	} = inputs;
 
 	const recipientIdFromPublicKey = recipientPublicKey
-		? cryptography.getAddressFromPublicKey(recipientPublicKey)
+		? getAddressFromPublicKey(recipientPublicKey)
 		: undefined;
 	const recipientId = inputs.recipientId
 		? inputs.recipientId
 		: recipientIdFromPublicKey;
 
-	const transaction: PartialTransaction = {
-		type: 0,
-		amount: amount.toString(),
+	const transaction = {
+		...createBaseTransaction(inputs),
+		asset: data ? { data } : {},
+		amount,
 		fee: TRANSFER_FEE.toString(),
-		recipientId,
+		recipientId: recipientId as string,
 		recipientPublicKey,
-		asset: createAsset(data),
+		type: 0,
 	};
 
-	return prepareTransaction(
-		transaction,
-		passphrase,
-		secondPassphrase,
-		timeOffset,
-	) as TransferTransaction;
+	if (!passphrase) {
+		return transaction;
+	}
+
+	const transactionWithSenderInfo = {
+		...transaction,
+		recipientId: recipientId as string,
+		senderId: transaction.senderId as string,
+		senderPublicKey: transaction.senderPublicKey as string,
+	};
+
+	const transferTransaction = new TransferTransaction(
+		transactionWithSenderInfo,
+	);
+	transferTransaction.sign(passphrase, secondPassphrase);
+
+	return transferTransaction.toJSON();
 };
