+++ packages/lisk-transactions/test/transactions/0_transfer.ts
@@ -23,7 +23,6 @@ import {
 	validTransferAccount,
 	validTransferTransactions,
 } from '../../fixtures';
-import * as utils from '../../src/utils';
 
 describe('Transfer transaction class', () => {
 	const validTransferTransaction = addTransactionFields(
@@ -65,166 +64,6 @@ describe('Transfer transaction class', () => {
 			expect(validSelfTransferTestTransaction.fee.toString()).to.eql(
 				TRANSFER_FEE.toString(),
 			);
-		});
-	});
-
-	describe('#create', () => {
-		const timeWithOffset = 38350076;
-		const passphrase = 'secret';
-		const secondPassphrase = 'second secret';
-		let result: object;
-
-		beforeEach(async () => {
-			sandbox.stub(utils, 'getTimeWithOffset').returns(timeWithOffset);
-		});
-
-		describe('when the transaction is created with one passphrase and asset data', () => {
-			beforeEach(async () => {
-				result = TransferTransaction.create({
-					passphrase,
-					amount: '1000',
-					recipientId: '1L',
-					data: 'tx info',
-				});
-			});
-
-			it('should create transfer transaction ', async () => {
-				expect(result).to.have.property('id');
-				expect(result).to.have.property('type', 0);
-				expect(result).to.have.property('amount', '1000');
-				expect(result).to.have.property('fee', TRANSFER_FEE.toString());
-				expect(result).to.have.property('senderId');
-				expect(result).to.have.property('senderPublicKey');
-				expect(result).to.have.property('recipientId', '1L');
-				expect(result).to.have.property('timestamp', timeWithOffset);
-				expect(result).to.have.property('signature').and.not.to.be.empty;
-				expect((result as any).asset.data).to.eql('tx info');
-			});
-
-			it('should use time.getTimeWithOffset to calculate the timestamp', async () => {
-				expect(utils.getTimeWithOffset).to.be.calledWithExactly(undefined);
-			});
-
-			it('should use time.getTimeWithOffset with an offset of -10 seconds to calculate the timestamp', async () => {
-				const offset = -10;
-				TransferTransaction.create({
-					passphrase,
-					amount: '1000',
-					recipientId: '1L',
-					data: 'a',
-					timeOffset: offset,
-				});
-				expect(utils.getTimeWithOffset).to.be.calledWithExactly(offset);
-			});
-		});
-
-		describe('when the transaction is created with first and second passphrase and no asset data', () => {
-			beforeEach(async () => {
-				result = TransferTransaction.create({
-					amount: '1000',
-					recipientId: '1L',
-					passphrase,
-					secondPassphrase,
-				});
-			});
-
-			it('should create transfer transaction ', async () => {
-				expect(result).to.have.property('id');
-				expect(result).to.have.property('type', 0);
-				expect(result).to.have.property('amount', '1000');
-				expect(result).to.have.property('fee', TRANSFER_FEE.toString());
-				expect(result).to.have.property('senderId');
-				expect(result).to.have.property('senderPublicKey');
-				expect(result).to.have.property('recipientId', '1L');
-				expect(result).to.have.property('timestamp', timeWithOffset);
-				expect(result).to.have.property('signature').and.not.to.be.empty;
-				expect(result).to.have.property('signSignature').and.not.to.be.empty;
-			});
-		});
-
-		describe('when the transaction is created with invalid inputs', () => {
-			it('should throw an invalid input error when data is not string', () => {
-				expect(
-					TransferTransaction.create.bind(undefined, {
-						passphrase,
-						amount: '1000',
-						recipientId: '18278674964748191682L',
-						data: 1 as any,
-					}),
-				).to.throw('Invalid asset types');
-			});
-
-			it('should throw an invalid input error when data is too long', () => {
-				expect(
-					TransferTransaction.create.bind(undefined, {
-						passphrase,
-						secondPassphrase,
-						amount: '1000',
-						recipientId: '18278674964748191682L',
-						data:
-							'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.',
-					}),
-				).to.throw('Transaction data field cannot exceed 64 bytes.');
-			});
-
-			it('should throw an invalid input error when amount is invalid', () => {
-				expect(
-					TransferTransaction.create.bind(undefined, {
-						passphrase,
-						secondPassphrase,
-						amount: 1000 as any,
-						recipientId: '18278674964748191682L',
-					}),
-				).to.throw('Amount must be a valid number in string format.');
-			});
-
-			it('should throw an invalid input error when recipientId not provided', () => {
-				expect(
-					TransferTransaction.create.bind(undefined, {
-						passphrase,
-						secondPassphrase,
-						amount: '1000',
-					}),
-				).to.throw('`recipientId` must be provided.');
-			});
-
-			it('should throw an invalid input error when recipientId does not match recipientPublicKey', () => {
-				expect(
-					TransferTransaction.create.bind(undefined, {
-						passphrase,
-						secondPassphrase,
-						amount: '1000',
-						recipientId: '18278674964748191682L',
-						recipientPublicKey:
-							'24193236b7cbeaf5e6feafbbf7a791095ea64ec73abde8f0470001fee5d39d9d',
-					}),
-				).to.throw('recipientId does not match recipientPublicKey.');
-			});
-		});
-
-		describe('when the transaction is created without passphrase', () => {
-			beforeEach(async () => {
-				result = TransferTransaction.create({
-					amount: '1000',
-					recipientId: '18278674964748191682L',
-				});
-			});
-
-			it('should create transfer transaction ', async () => {
-				expect(result).to.have.property('type', 0);
-				expect(result).to.have.property('amount', '1000');
-				expect(result).to.have.property('fee', TRANSFER_FEE.toString());
-				expect(result)
-					.to.have.property('timestamp')
-					.and.equal(timeWithOffset);
-				expect((result as any).senderPublicKey).to.be.undefined;
-				expect(result)
-					.to.have.property('recipientId')
-					.and.to.equal('18278674964748191682L');
-				expect(result).not.to.have.property('id');
-				expect(result).not.to.have.property('signature');
-				expect(result).not.to.have.property('signSignature');
-			});
 		});
 	});
